+++ CHANGELOG.md
@@ -21,6 +21,7 @@
     -   [is_scalar](./docs/variables/is_scalar.md)
     -   [is_string](./docs/variables/is_string.md)
     -   [is_symbol](./docs/variables/is_symbol.md)
+    -   [is_undefined](./docs/variables/is_undefined.md)
 
 -   Добавлена функция [array](./docs/array/array.md) в модуль [array](./docs/array.md)